Refactor: StorageStub to separate class
Move StorageStub from CommandTestUtil to seedu.mark.storage in test
module, to match ModelStub.
