+++ src/test/java/seedu/mark/logic/commands/AddCommandIntegrationTest.java
@@ -7,11 +7,11 @@ import static seedu.mark.testutil.TypicalBookmarks.getTypicalMark;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import seedu.mark.logic.commands.CommandTestUtil.StorageStub;
 import seedu.mark.model.Model;
 import seedu.mark.model.ModelManager;
 import seedu.mark.model.UserPrefs;
 import seedu.mark.model.bookmark.Bookmark;
+import seedu.mark.storage.StorageStub;
 import seedu.mark.testutil.BookmarkBuilder;
 
 /**
